// src/interface/types.ts
export interface Coverage {
  ao: number | boolean;
  bonusProtection: number | boolean;
  glassProtection: number | boolean;
}

export interface Discounts {
  voucher: number | boolean;
  adviserDiscount: number | boolean;
  carSurcharge: number | boolean;
  commercialDiscount: number | boolean;
  vipDiscount: number | boolean;
}